Remove legacy kmstool
Kmstool started off with a version that implemented KMS TLS termination
in the parent instance. This is an anti pattern, because it does not
provide authentication between enclave and KMS.

To not confuse users of Nitro Enclaves with example code that follows
bad practices, this patch removes all traces of the old kmstool.

Signed-off-by: Alexander Graf <graf@amazon.com>